+++ docs/CHANGE_LOG.md
@@ -2,11 +2,15 @@
 
 ## Version 0.5.0:
 
+#### **New Features:**
+- Added specific properties watch support! This will only watch for changes on specific properties on an object, instead of all of them by default.
+
 #### **Binding Options / Function Options - Custom Triggers:**
 - Added a new binding/option custom trigger called "onRemove", which states an event that should be triggered when a DOM element is removed.
 
 #### **Public Functions:**
 - Added new public function "resumeWatch()", which is used to resume a watch that has been paused.
+- Added new parameter "propertyNames" to the "watch()" public function, which states the property names to watch changes for (defaults to all).
 
 #### **Documentation:**
 - Fixed some of the documentation mistakes.